Add unit tests and fix broken ones
