+++ src/components/DataEntry/tests/DataEntry.test.tsx
@@ -1,55 +1,11 @@
 import React from "react";
-
-import GoalSelectorScroll from "..";
-
+import ReactDOM from "react-dom";
 import { Provider } from "react-redux";
-import { act, Simulate } from "react-dom/test-utils";
-import renderer, {
-  ReactTestInstance,
-  ReactTestRenderer
-} from "react-test-renderer";
-import AddWords from "..";
-import configureStore from "redux-mock-store";
-import ConnectedDataEntry, { DataEntry } from "../DataEntryComponent";
-import { LocalizeProvider } from "react-localize-redux";
 import configureMockStore from "redux-mock-store";
-import createMockStore from "redux-mock-store";
-import { GoalSelectorState } from "../../../types/goals";
-import axios from "axios";
-import { Word } from "../../../types/word";
+import ConnectedDataEntryComponent from "../";
+import LocalizedDataEntryComponent from "../DataEntryComponent";
+import SemanticDomain from "../../TreeView/SemanticDomain";
 
-jest.mock("@material-ui/core", () => {
-  const material = jest.requireActual("@material-ui/core");
-  return {
-    ...material,
-    Dialog: material.Box
-  };
-});
-
-let master: renderer.ReactTestRenderer;
-let handle: renderer.ReactTestInstance;
-
-const testWord: Word = {
-  id: "",
-  vernacular: "",
-  senses: [
-    {
-      glosses: [],
-      semanticDomains: []
-    }
-  ],
-  audio: "",
-  created: "",
-  modified: "",
-  history: [],
-  partOfSpeech: "",
-  editedBy: [],
-  accessability: 0,
-  otherField: "",
-  plural: ""
-};
-
-// Mock getTranslate
 const MOCK_TRANSLATE = jest.fn(_ => {
   return "dummy";
 });
@@ -62,157 +18,47 @@ jest.mock("react-localize-redux", () => {
     })
   };
 });
-jest.mock("axios");
-let mockedAxios = axios as jest.Mocked<typeof axios>;
 
-// Circumvent unneeded store connections
-jest.mock("../../TreeView", () => {
-  const material = jest.requireActual("@material-ui/core");
-  return material.Container;
-});
+const createMockStore = configureMockStore([]);
 
-// Mock store
-const mockStore = configureMockStore()({
-  treeViewState: { currentDomain: { name: "en", number: "1", subDomains: [] } }
-});
+const mockSemanticDomain: SemanticDomain = {
+  name: "",
+  id: "",
+  subdomains: []
+};
 
-beforeEach(() => {
-  // Here, use the act block to be able to render our AddWords into the DOM
-  // Re-created each time to prevent actions from previous runs from affecting future runs
-  act(() => {
-    // Getfrontierwords
-    mockedAxios.get.mockImplementationOnce(url => {
-      return Promise.resolve({ data: [] });
-    });
-    master = renderer.create(
-      <ConnectedDataEntry domain={{ name: "en", id: "1", subdomains: [] }} />
+const mockState = {
+  localize: {
+    languages: [],
+    translations: null,
+    options: null
+  },
+  treeViewState: {
+    currentdomain: mockSemanticDomain
+  }
+};
+
+const mockStore = createMockStore(mockState);
+
+describe("Tests DataEntryComponent", () => {
+  it("connected data entry component renders without crashing", () => {
+    const div = document.createElement("div");
+    ReactDOM.render(
+      <Provider store={mockStore}>
+        <ConnectedDataEntryComponent />
+      </Provider>,
+      div
     );
+    ReactDOM.unmountComponentAtNode(div);
   });
-  handle = master.root.findByType(DataEntry);
-
-  mockedAxios.put.mockClear();
-});
-
-afterAll(() => {
-  jest.unmock("../../TreeView");
-});
-
-describe("Tests AddWords", () => {
-  it("Constructs correctly", () => {
-    // Switch from the selectDomain view to normal view
-    handle.instance.setState({
-      ...handle.instance.state,
-      gettingSemanticDomain: false
-    });
-    snapTest("default view");
-  });
-
-  it("Adds a word", done => {
-    handle.instance.setState({ newVern: "testVern", newGloss: "testGloss" });
-    mockedAxios.post.mockImplementationOnce((url, word: Word) => {
-      return Promise.resolve({ data: "123" });
-    });
-    handle.instance.submit(undefined, () => {
-      expect(handle.instance.state.rows).toEqual([
-        {
-          vernacular: "testVern",
-          glosses: "testGloss",
-          id: "123",
-          senseIndex: 0,
-          dupId: ""
-        }
-      ]);
-      done();
-    });
-  });
-
-  it("Edits a word", done => {
-    handle.instance.setState({
-      rows: [
-        {
-          vernacular: "testVern1",
-          glosses: "testGloss1",
-          id: "123",
-          senseIndex: 0,
-          dupId: ""
-        },
-        {
-          vernacular: "testVern2",
-          glosses: "testGloss2",
-          id: "456",
-          senseIndex: 0,
-          dupId: ""
-        },
-        {
-          vernacular: "testVern3",
-          glosses: "testGloss3",
-          id: "789",
-          senseIndex: 0,
-          dupId: ""
-        }
-      ]
-    });
-    mockedAxios.get.mockResolvedValueOnce({ data: testWord });
-    mockedAxios.put.mockResolvedValueOnce({ data: 1 });
-    handle.instance.updateWord(1, () => {
-      expect(mockedAxios.put).toHaveBeenCalledTimes(1);
-      done();
-    });
-  });
-
-  it("Removes a word", done => {
-    handle.instance.setState({
-      rows: [
-        {
-          vernacular: "testVern1",
-          glosses: "testGloss1",
-          id: "123",
-          senseIndex: 0,
-          dupId: ""
-        },
-        {
-          vernacular: "testVern2",
-          glosses: "testGloss2",
-          id: "456",
-          senseIndex: 0,
-          dupId: ""
-        },
-        {
-          vernacular: "testVern3",
-          glosses: "testGloss3",
-          id: "789",
-          senseIndex: 0,
-          dupId: ""
-        }
-      ]
-    });
-    mockedAxios.delete.mockResolvedValue(1);
-    handle.instance.removeWord(456, () => {
-      handle.instance.removeRow(1);
-      expect(handle.instance.state.rows).toEqual([
-        {
-          vernacular: "testVern1",
-          glosses: "testGloss1",
-          id: "123",
-          senseIndex: 0,
-          dupId: ""
-        },
-        {
-          vernacular: "testVern3",
-          glosses: "testGloss3",
-          id: "789",
-          senseIndex: 0,
-          dupId: ""
-        }
-      ]);
-      done();
-    });
+  it("localized data entry component renders without crashing", () => {
+    const div = document.createElement("div");
+    ReactDOM.render(
+      <Provider store={mockStore}>
+        <LocalizedDataEntryComponent domain={mockSemanticDomain} />
+      </Provider>,
+      div
+    );
+    ReactDOM.unmountComponentAtNode(div);
   });
 });
-
-// Utility functions -----------------------------
-
-// Perform a snapshot test
-function snapTest(name: string) {
-  expect(master.toJSON()).toMatchSnapshot();
-}
